+++ src/domain/inconsistent_change.go
@@ -1,29 +1,8 @@
 package domain
 
-type InconsistentChanges []InconsistentChange
-
-func (ic InconsistentChanges) BranchNames() LocalBranchNames {
-	result := make(LocalBranchNames, len(ic))
-	for i, change := range ic {
-		result[i] = change.Before.LocalName
-	}
-	return result
-}
-
 // InconsistentChange describes a change where both local and remote branch exist before and after,
 // but it's not an OmniChange, i.e. the SHA are different.
 type InconsistentChange struct {
 	Before BranchInfo
 	After  BranchInfo
-}
-
-func (ic InconsistentChanges) Categorize(branchTypes BranchTypes) (perennials, features InconsistentChanges) {
-	for _, change := range ic {
-		if branchTypes.IsFeatureBranch(change.Before.LocalName) {
-			features = append(features, change)
-		} else {
-			perennials = append(perennials, change)
-		}
-	}
-	return
 }
